Refactor
Altera o tipo de string para string? e adiciona um validações nos nomes, acabando com a nulidade.

+++ Crud_biblioteca/Model/Livro.cs
@@ -5,22 +5,22 @@
         public int Id { get; set; }
         public string Nome { get; set; } = "Não informado";
         public int Quantidade { get; set; }
-        public double Valor { get; set; }
+        public decimal Valor { get; set; }
 
-        public Livro(int id, string nome, int quantidade, double valor)
+        public Livro(int id, string nome, int quantidade, decimal valor)
         {
             Id = id;
             Nome = nome;
             Quantidade = quantidade;
             Valor = valor;
         }
-        public Livro() 
-        { 
-        
+        public Livro()
+        {
+
         }
 
-        public override string ToString() 
-        { 
+        public override string ToString()
+        {
             return $"Id: {Id} - Nome: {Nome} - Quantidade: {Quantidade} - Valor: {Valor}";
         }
